feat: add Landscape area polygon to the map
This change updates the Map component to support it receiving a GeoJSON
objects. This GeoJSON is used to add a new map polygon layer.

The new `geojson` property is already being used to the Landscape's
`areaPolygon` attribute. When a `areaPolygon` is used, the bounding box
of this polygon is used. Otherwise, the Landscape's position bounding
box continue being used.

Even before this change the `LandscapeView` had a limitation to don't
properly handle the situation where a Landscape has no location. This
limitation isn't addressed in this change.