feat: add tracking to landscape creation when outline provided
This doesn't track the case when the user skips providing the outline,
will clarify if that is needed or not (maybe we are only interested in
when the outline is present). Also this step will be repeated whenever
the boundary is changed. Also it does not include the country yet.

I tried a few different solutions, but I've settled here for a simple one.