feat: Merge plausible changes to main (#586)
* feat: Track when users click on KoboToolbox link (#488)

Simply changed the Tool component to use the ExternalLink component. I
think it is fair to expect all tools to be external links here.

* feat: add tracking to landscape creation when outline provided

This doesn't track the case when the user skips providing the outline,
will clarify if that is needed or not (maybe we are only interested in
when the outline is present). Also this step will be repeated whenever
the boundary is changed. Also it does not include the country yet.

I tried a few different solutions, but I've settled here for a simple one.

* feat: wait for save to complete in landscape creation tracking

* feat: add country to landscape creation tracking payload

This will track:
- the country code, not the name
- will be undefined if the country was not previously set

* feat: track skipping boundary upload with landscape creation

Pretty similar to the last two commits, just had to find out where the
skip boundary landscape was being saved.

* feat: only track Landscape boundary changes on creation (#490)

* feat: make save functions async (#518)

This means that we can wait for them to finish before we send a
tracking event.

* feat: track group creation (#490)

* feat: track account creation after signup (#492)

This uses a cookie from the backend to detect that the user recently
created an account. If the cookie is set, the tracking event is sent
and the cookie is deleted.

* Revert "feat: track account creation after signup (#492)"

This reverts commit 2703c0380dbcb7bba3866803ae7f79696a2f484f.

The commit is no longer necessary because we will handle the user
signup tracking entirely on the backend.

* feat: add dataset ID (#491)

* feat: Use variable for all events on boundary save (#518)