fix: Display null columns in Mapping data preview (#633)
* fix: Show null columns in dataset preview

The problem: `Sheets.utils.sheet_to_json`, when called with '{header:
1}' will leave blank spots in the array. For example: `1,,3` becomes
`Array(3)[1, <1 empty slot>, 3]`. This is a Javascript array
quirk. However the previous code iterated over an array, to these
blank columns were just "skipped".

This commit changes `Sheets.utils.sheet_to_json` to use `{header:
headers}`. Because the headers are parsed from the sheet itself, this
should preserve the order of the heaers as well.

* chore: Add basic test for data preview

* chore: Test null columns still rendered