test: mock useIsMounted to prevent intermittently failing tests
When determining the place information by name, `useIsMounted` is called to
determine if the boundingBox must be set.
See: src/landscape/components/LandscapeForm/BoundaryStep/index.js

During tests, it was found that `useIsMounted` returns sometimes true,
sometimes false. This causes tests to sometimes uses the response from
`getPlaceInfoByName`, and sometimes to not.

During tests, it seemed that the response should always be the non-mounted one.
Mocking `useIsMounted` enforces this for all tests.