fix: Plausible update, analytics updates for the Share Public Link feature #533 (#1631)
* feat: standardize event tracking for link sharing

add a common analytics event for sharing and reuse it in both copy-link share button usages in the application. add a test for the new common code.

* feat: Standardize download analytics events, add location value

Add another event to the events.js file and use it in both list-view and standalone download pages. Add testing where relevant and do some cleanup on the prior feature's implementation.

* fix: change div to self-closing in unit test component

Co-authored-by: Paul Schreiber <paul@techmatters.org>

* fix: update copyright date

Co-authored-by: Paul Schreiber <paul@techmatters.org>

* fix: update copyright date

Co-authored-by: Paul Schreiber <paul@techmatters.org>

* fix: use conventional property name for download location

---------

Co-authored-by: Paul Schreiber <paul@techmatters.org>